blueprint/requirements.txt
# ============================================================
# TỰ MINH AGI - BLUEPRINT SYSTEM DEPENDENCIES
# ============================================================

# Core dependencies
pyyaml>=6.0.1
asyncio

# Ollama integration
ollama>=0.1.7

# RAG & Vector DB
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Web Search
duckduckgo-search>=4.1.1

# NLP & Embeddings
transformers>=4.36.0
torch>=2.1.0

# UI (Gradio)
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1

# Logging & Monitoring
loguru>=0.7.2

# PDF Processing (optional)
pypdf2>=3.0.1
pdfplumber>=0.10.3

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Code Analysis (optional)
pylint>=3.0.3
black>=23.12.0
